Move DeploymentManager back into test dir since only used in integrations package
