+++ contracts/integrations/src/index.ts
@@ -1,4 +1,3 @@
 export * from './artifacts';
 export * from './wrappers';
 export * from './function_assertions';
-export { DeploymentManager } from './deployment_manager';
