+++ contracts/integrations/test/coordinator/deploy_coordinator.ts
@@ -3,7 +3,7 @@ import { artifacts as exchangeArtifacts } from '@0x/contracts-exchange';
 import { BlockchainTestsEnvironment } from '@0x/contracts-test-utils';
 import { BigNumber } from '@0x/utils';
 
-import { DeploymentManager } from '../../src/deployment_manager';
+import { DeploymentManager } from '../deployment_manager';
 
 /**
  * Deploys a Coordinator contract configured to work alongside the provided `deployment`.
