+++ contracts/integrations/test/forwarder/forwarder_test.ts
@@ -20,8 +20,8 @@ import {
 import { ForwarderRevertErrors } from '@0x/order-utils';
 import { BigNumber } from '@0x/utils';
 
-import { DeploymentManager } from '../../src/deployment_manager';
 import { Actor, actorAddressesByName, FeeRecipient, Maker } from '../actors';
+import { DeploymentManager } from '../deployment_manager';
 
 import { deployForwarderAsync } from './deploy_forwarder';
 import { ForwarderTestFactory } from './forwarder_test_factory';
