+++ contracts/integrations/test/internal-integration-tests/deployment_test.ts
@@ -33,7 +33,7 @@ import { AssetProxyId } from '@0x/types';
 import { BigNumber } from '@0x/utils';
 import { TxData } from 'ethereum-types';
 
-import { AssetProxyDispatcher, Authorizable, Ownable } from '../../src/wrapper_interfaces';
+import { AssetProxyDispatcher, Authorizable, Ownable } from '../wrapper_interfaces';
 
 // tslint:disable:no-unnecessary-type-assertion
 blockchainTests('Deployment and Configuration End to End Tests', env => {
